requirement.txt
transformers==4.31.0
datasets
deepspeed==0.14.5
sentencepiece
tensorboardX
accelerate==0.31.0

einops
importlib-resources
jsonlines
numexpr
omegaconf
pycountry
pytablewriter
rouge-score
sacrebleu
scikit-learn
sqlitedict
tqdm-multiprocess
zstandard

bitsandbytes
peft==0.13.1
fraction
wandb
matplotlib